requirements.txt
pillow-heif
pillow
pysimplegui